+++ dojo/product_type/__init__.py
@@ -1,0 +1,1 @@
+import dojo.product_type.admin  # noqa: F401
