+++ dojo/product_type/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+from dojo.product_type.models import Product_Type
+
+
+@admin.register(Product_Type)
+class Product_TypeAdmin(admin.ModelAdmin):
+
+    """Admin support for the Product_Type model."""
